requirements.txt
# Core dependencies
openai>=1.51.0
openai-agents>=0.3.0
python-dotenv>=1.0.0
httpx>=0.24.0

# Optional dependencies for enhanced functionality
pydantic>=2.0.0
aiohttp>=3.8.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0